fixed some responsible party handling
in short: for new patients who are their own responsible party, there is a NULL
in the resparty field. this means that the patient will not show up in their own
family display. BUMMER! so added a test in update_pt() to put in the id of the
new patient if their resparty is NULL.