added some torture testing
basically a way to add a gazillion patients, appointments, and
treatments

trying to see how the database would hold up ... was thinking I would
need an index on journal ts, but with 1448 entries in the journal it
seems to be working just fine ... but I should probably still have the
index, in particular as the number of appointments and journal entries
mounts, and it will grow more or less without bounds

in fact, I need to plan ahead. what to do when the database gets
ginormous? how big will it get? my test database was about 13k, and
the torture database with 413 patients, 1448 appointments and 5349
treatments is only 388k. but ...